reaname auth library
