+++ src/util/GoogleAuth/index.js
@@ -1,0 +1,7 @@
+import GoogleSignIn from './GoogleSignIn'
+import useGoogleSignIn from './useGoogleSignIn'
+import { withGoogleAuth } from './withGoogleAuth'
+
+export { GoogleSignIn, useGoogleSignIn, withGoogleAuth }
+
+export default useGoogleSignIn
