--- src/util/withGoogleAuth/index.js
@@ -1,3 +1,0 @@
-import { withGoogleAuth } from './withGoogleAuth'
-
-export default withGoogleAuth
